refactor: 피드백 반영(책임 분리, range)
