+++ src/main/kotlin/minesweeper/domain/board/MineSweeperBoard.kt
@@ -1,7 +1,7 @@
 package minesweeper.domain.board
 
 import minesweeper.domain.position.MineSweeperPositions
-import minesweeper.domain.position.Position.Companion.MIN_POSITION
+import minesweeper.domain.position.Positions
 
 class MineSweeperBoard(
     boardRange: BoardRange,
@@ -14,9 +14,12 @@ class MineSweeperBoard(
         require(mineQuantity >= MIN_MINE_QUANTITY) { "지뢰 개수는 $MIN_MINE_QUANTITY 개 이상이어야 합니다." }
         require(mineQuantity <= boardRange.calculateArea()) { "지뢰 개수는 전체 크기를 초과할 수 없습니다." }
 
-        val minePositions = boardRange.createRandomPositions(mineQuantity)
-        val widthPositionRange = IntRange(MIN_POSITION, boardRange.maxWidth())
-        val heightPositionRange = IntRange(MIN_POSITION, boardRange.maxHeight())
+        val minePositions = Positions.createRandomPositions(
+            minRandomPositionSize = mineQuantity,
+            boardRange = boardRange,
+        )
+        val widthPositionRange = boardRange.widthRange()
+        val heightPositionRange = boardRange.heightRange()
         board = heightPositionRange.map { yPosition ->
             MineSweeperPositions.create(widthPositionRange, yPosition, minePositions)
         }
